Change the Expression parser to use doubles instead of ints.
git-svn-id: file:///srv/svn/repo/trunk/Programs/Compiler@306 c9e57fc7-f7bf-4014-a52f-2ff4955da382

+++ src/test/parser/Expression.java
@@ -29,7 +29,7 @@ public class Expression
     return sum.toString();
   }
   
-  public int evaluate()
+  public double evaluate()
   {
     return sum.evaluate();
   }
